Log system finished
